Round 30: give UDP a live caller + fix a silent heap-canary ABI bug
Feature (per feature-ledger priority #2): the UDP socket API
(socket_udp_send/socket_udp_recv, Phase 6's socket-style API) had run
only inside udp_self_test's hand-fed traffic since round 21 added the
ARP-cache-hit requirement to socket_udp_send. Added a shell "udpecho
<text>" command: a self-talk-over-loopback round trip to 127.0.0.1,
pre-populating the ARP cache for self (matching udp_self_test's own
established precedent for the same requirement), then a small inlined
echo-service loop built directly on socket_udp_recv's own reported
out_meta (dst_port match), auto-responding the same "no separate
caller involvement" way icmp_poll/tcp_conn_poll's auto-responders
already do -- no new *_poll helper needed since UDP has no per-
connection state for another caller to share. Worked correctly on the
first try. Extended task_e's existing shared-netif-queue concurrency
comment to name this as a third consumer, same already-accepted
category as ping/tcpecho. Formalized into phase4_milestone.py in this
same round, verified reliable across 2 runs.

Adding udpecho's locals to shell_dispatch pushed task_f's own
#[bounded_stack(bytes=2048)] past its budget -- compiler-reported
worst case 2184 bytes via task_f -> shell_dispatch -> tcp_conn_poll ->
... -> buf_read_u16_be. Raised to 4096 (real compile-time catch, same
class as round 27's task_e fix).

Bugfix (found via live-testing scrutiny while verifying the above, not
via udpecho itself): heap_usage_self_test has silently printed "(FAIL,
expect headroom >= 16384)" at every boot since round 10 introduced it
-- confirmed present on round 29's own already-committed baseline, so
not a round-30 regression. Root cause: dhruva_heap_used_bytes returned
plain `long`, 32-bit under this target's AAPCS (arm-none-eabi ILP32),
but vani's own extern declaration says `-> i64`, a genuine 64-bit
return. A 32-bit `long` only populates r0 of the r0:r1 AAPCS return
pair, leaving r1 as leftover garbage from a prior call -- the reported
"used=572811198400528" decoded to a perfectly reasonable 67600 in its
low 32 bits, with non-deterministic garbage in the high 32. Fixed by
returning `long long` end to end; now reads "used=67600 ... (PASS)".

Verified: full self-test battery unchanged and now genuinely 0 FAIL
(was silently 1), phase4_milestone.py PASS across 2 runs including the
new udpecho step, USB enumeration unchanged with both usb-net/
usb-storage device types (1 boot banner each), heap_stress.py PASS (1
boot banner), power_yank.py PASS (54/54).

+++ boot/rpi1/runtime_stubs.c
@@ -163,7 +163,27 @@ void *dhruva_alloc_bytes(long n) {
  * early-warning canary added by the same round-10 fix that resized
  * this heap, so a future round eating back into the new headroom
  * fails loudly at boot instead of reproducing this exact bug again
- * as a silent reboot loop. */
-long dhruva_heap_used_bytes(void) {
-    return (long)dhruva_heap_used;
+ * as a silent reboot loop.
+ *
+ * BUG (round 30 audit, found via live-testing scrutiny of the boot
+ * log, not this self-test's own PASS/FAIL count -- the same "read the
+ * actual output, don't just trust a name" discipline that caught
+ * round 26's ARP-cache bug): this returned plain `long`, which is
+ * 32-bit under this target's AAPCS (arm-none-eabi ILP32), but vani's
+ * own extern declaration (kernel_main.vani line ~33) says `-> i64`, a
+ * genuine 64-bit return. AAPCS returns a 64-bit value in the r0:r1
+ * register pair; a 32-bit `long` return only ever populates r0,
+ * leaving r1 as whatever was left over from a prior call. Every
+ * caller reading this as a real i64 got a huge, non-deterministic
+ * garbage value in the high 32 bits (heap_usage_self_test's own boot
+ * print showed something like "used=572811198400528" instead of the
+ * real, entirely reasonable ~67600) -- silently printing "(FAIL,
+ * expect headroom >= 16384)" at every single boot since this function
+ * was introduced (round 10), never noticed because attention went to
+ * self-tests' PASS/FAIL counts, not this one line's own literal text.
+ * Fixed by actually returning a 64-bit value end to end, matching
+ * vani's declared i64 rather than silently truncating across the FFI
+ * boundary. */
+long long dhruva_heap_used_bytes(void) {
+    return (long long)dhruva_heap_used;
 }
